+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Fix returning values from `Object.prototype`
+
 2.0.0 / 2020-04-19
 ==================